Move tests and logic to sub folders to accommodate growth
